requirements.txt
streamlit
python-dotenv
langchain
faiss-cpu
pypdf
huggingface_hub
sentence-transformers
together
